refactor(aria/combobox): deprecate and relocate legacy combobox entrypoints
Relocate the public `src/aria/combobox` entrypoint to `src/aria/deprecated/combobox` and the private implementation patterns to `src/aria/private/deprecated/combobox`. This frees up the primary `combobox` namespace for the upcoming standalone combobox widget migration.

- Add JSDoc `@deprecated` and `@breaking-change 22.0.0` markers to legacy classes and primitives.
- Adjust relative imports outwards to accommodate the deeper directory nesting levels.
- Re-wire build dependencies and entrypoint configurations across the framework package graph.
- Bulk update consumer example module targets and prune unused dependencies.
