+++ goldens/aria/deprecated/combobox/index.api.md
@@ -1,4 +1,4 @@
-## API Report File for "@angular/aria_combobox"
+## API Report File for "@angular/aria_deprecated_combobox"
 
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
@@ -8,7 +8,7 @@ import * as _angular_cdk_bidi from '@angular/cdk/bidi';
 import * as _angular_core from '@angular/core';
 import { OnDestroy } from '@angular/core';
 
-// @public
+// @public @deprecated
 export class Combobox<V> {
     constructor();
     readonly alwaysExpanded: _angular_core.InputSignalWithTransform<boolean, unknown>;
@@ -30,7 +30,7 @@ export class Combobox<V> {
     static ɵfac: _angular_core.ɵɵFactoryDeclaration<Combobox<any>, never>;
 }
 
-// @public
+// @public @deprecated
 export class ComboboxDialog {
     constructor();
     // (undocumented)
@@ -46,7 +46,7 @@ export class ComboboxDialog {
     static ɵfac: _angular_core.ɵɵFactoryDeclaration<ComboboxDialog, never>;
 }
 
-// @public
+// @public @deprecated
 export class ComboboxInput {
     constructor();
     readonly combobox: Combobox<any>;
@@ -58,7 +58,7 @@ export class ComboboxInput {
     static ɵfac: _angular_core.ɵɵFactoryDeclaration<ComboboxInput, never>;
 }
 
-// @public
+// @public @deprecated
 export class ComboboxPopup<V> {
     readonly combobox: Combobox<V> | null;
     readonly _controls: _angular_core.WritableSignal<ComboboxListboxControls<any, V> | ComboboxTreeControls<any, V> | ComboboxDialogPattern | undefined>;
@@ -68,7 +68,7 @@ export class ComboboxPopup<V> {
     static ɵfac: _angular_core.ɵɵFactoryDeclaration<ComboboxPopup<any>, never>;
 }
 
-// @public
+// @public @deprecated
 export class ComboboxPopupContainer {
     // (undocumented)
     static ɵdir: _angular_core.ɵɵDirectiveDeclaration<ComboboxPopupContainer, "ng-template[ngComboboxPopupContainer]", ["ngComboboxPopupContainer"], {}, {}, never, never, true, [{ directive: typeof DeferredContent; inputs: {}; outputs: {}; }]>;
