+++ src/aria/listbox/public-api.ts
@@ -17,4 +17,4 @@ export {
   ComboboxInput as ɵɵComboboxInput,
   ComboboxPopup as ɵɵComboboxPopup,
   ComboboxPopupContainer as ɵɵComboboxPopupContainer,
-} from '../combobox';
+} from '../deprecated/combobox';
