+++ src/aria/private/BUILD.bazel
@@ -19,8 +19,8 @@ ts_project(
         "//:node_modules/@angular/core",
         "//src/aria/private/accordion",
         "//src/aria/private/behaviors/signal-like",
-        "//src/aria/private/combobox",
         "//src/aria/private/deferred-content",
+        "//src/aria/private/deprecated/combobox",
         "//src/aria/private/grid",
         "//src/aria/private/listbox",
         "//src/aria/private/menu",
