+++ src/aria/private/listbox/BUILD.bazel
@@ -13,7 +13,7 @@ ts_project(
         "//src/aria/private/behaviors/event-manager",
         "//src/aria/private/behaviors/list",
         "//src/aria/private/behaviors/signal-like",
-        "//src/aria/private/combobox",
+        "//src/aria/private/deprecated/combobox",
     ],
 )
