+++ src/aria/private/tree/BUILD.bazel
@@ -14,7 +14,7 @@ ts_project(
         "//src/aria/private/behaviors/expansion",
         "//src/aria/private/behaviors/signal-like",
         "//src/aria/private/behaviors/tree",
-        "//src/aria/private/combobox",
+        "//src/aria/private/deprecated/combobox",
     ],
 )
