+++ src/aria/tree/BUILD.bazel
@@ -10,7 +10,7 @@ ng_project(
     ),
     deps = [
         "//:node_modules/@angular/core",
-        "//src/aria/combobox",
+        "//src/aria/deprecated/combobox",
         "//src/aria/private",
         "//src/cdk/a11y",
         "//src/cdk/bidi",
