+++ src/components-examples/aria/aria-toolbar-simple-combobox/BUILD.bazel
@@ -12,7 +12,6 @@ ng_project(
     deps = [
         "//:node_modules/@angular/core",
         "//:node_modules/@angular/forms",
-        "//src/aria/combobox",
         "//src/aria/listbox",
         "//src/aria/simple-combobox",
         "//src/aria/toolbar",
