+++ src/components-examples/aria/autocomplete/autocomplete-auto-select/autocomplete-auto-select-example.ts
@@ -11,7 +11,7 @@ import {
   ComboboxInput,
   ComboboxPopup,
   ComboboxPopupContainer,
-} from '@angular/aria/combobox';
+} from '@angular/aria/deprecated/combobox';
 import {Listbox, Option} from '@angular/aria/listbox';
 import {
   afterRenderEffect,
