+++ src/components-examples/aria/combobox/combobox-dialog/combobox-dialog-example.ts
@@ -11,7 +11,7 @@ import {
   ComboboxDialog,
   ComboboxInput,
   ComboboxPopupContainer,
-} from '@angular/aria/combobox';
+} from '@angular/aria/deprecated/combobox';
 import {Listbox, Option} from '@angular/aria/listbox';
 import {
   afterRenderEffect,
