+++ src/components-examples/aria/select/BUILD.bazel
@@ -11,7 +11,7 @@ ng_project(
     ]),
     deps = [
         "//:node_modules/@angular/core",
-        "//src/aria/combobox",
+        "//src/aria/deprecated/combobox",
         "//src/aria/listbox",
         "//src/cdk/overlay",
     ],
