+++ src/components-examples/aria/select/select-disabled/select-disabled-example.ts
@@ -11,7 +11,7 @@ import {
   ComboboxInput,
   ComboboxPopup,
   ComboboxPopupContainer,
-} from '@angular/aria/combobox';
+} from '@angular/aria/deprecated/combobox';
 import {Listbox, Option} from '@angular/aria/listbox';
 import {ChangeDetectionStrategy, Component} from '@angular/core';
 import {OverlayModule} from '@angular/cdk/overlay';
